Added shared lib
